@@ -63,7 +63,7 @@ public class SecurityConfig {
 							  .addFilter(new CustomLoginFilter(authenticationManager));
 						
 		//api로 시작하는 요청에만 jwt필터 실행됨
-		http.requestMatchers().antMatchers("/api/customer/**")
+		http.requestMatchers().antMatchers("/api/question/**")
 							  .antMatchers("/api/engineer/**")
 							  .antMatchers("/api/admin/**")
 							  .and()
