@@ -87,7 +87,11 @@ public class UserController {
 	}
 
 	
-	
+	@GetMapping("/mypage")
+	public String myPage() {
+
+		return "/user/mypage";
+	}
 
 
 
