@@ -1,19 +1,38 @@
 package com.project.stms.service.s3;
 
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStream;
+import java.nio.file.Paths;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import software.amazon.awssdk.core.ResponseBytes;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.model.GetObjectRequest;
+import software.amazon.awssdk.services.s3.model.GetObjectResponse;
+import software.amazon.awssdk.services.s3.model.HeadObjectRequest;
+import software.amazon.awssdk.services.s3.model.HeadObjectResponse;
 import software.amazon.awssdk.services.s3.model.ListBucketsRequest;
 import software.amazon.awssdk.services.s3.model.ListBucketsResponse;
+import software.amazon.awssdk.services.s3.model.ListObjectsRequest;
+import software.amazon.awssdk.services.s3.model.ListObjectsResponse;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
 import software.amazon.awssdk.services.s3.model.PutObjectResponse;
 import software.amazon.awssdk.services.s3.model.S3Exception;
+import software.amazon.awssdk.services.s3.model.S3Object;
+import software.amazon.awssdk.transfer.s3.S3TransferManager;
+import software.amazon.awssdk.transfer.s3.model.CompletedFileDownload;
+import software.amazon.awssdk.transfer.s3.model.DownloadFileRequest;
+import software.amazon.awssdk.transfer.s3.model.FileDownload;
+import software.amazon.awssdk.transfer.s3.progress.LoggingTransferListener;
 
 @Component
 public class S3Service {
@@ -42,9 +61,9 @@ public class S3Service {
 			PutObjectRequest putOb = PutObjectRequest.builder().bucket(bucketName) // 버킷이름
 					.key(originName) // 파일이름
 					.metadata(metadata).build();
-			
+
 			PutObjectResponse response = s3.putObject(putOb, RequestBody.fromBytes(fileData));
-			
+
 			System.out.println("Successfully placed " + originName + " into bucket " + bucketName);
 
 			System.out.println(response.sdkHttpResponse().statusCode());
@@ -52,6 +71,86 @@ public class S3Service {
 		} catch (S3Exception e) {
 			System.err.println(e.getMessage());
 		}
+
+	}
+
+	public boolean readFiles(String keyName) {
+		
+		System.out.println("키네임 : " + keyName);
+		try {
+			ListObjectsRequest listObjects = ListObjectsRequest.builder().bucket(bucketName).build();
+
+			ListObjectsResponse res = s3.listObjects(listObjects);
+			List<S3Object> objects = res.contents();
+			for (S3Object myValue : objects) {
+
+				
+				System.out.print("\n" + myValue.key().equals( keyName));
+				System.out.print("\n The name of the key is " + myValue.key());
+				System.out.print("\n The owner is " + myValue.owner());
+				if(keyName.equals(myValue.key())) {
+					return true;
+				}
+			}
+
+		} catch (S3Exception e) {
+			System.err.println(e.awsErrorDetails().errorMessage());
+		}
+		
+		
+		return false;
+	}
+	
+	
+	public void checkFile(String keyName) {
+		
+		 try {
+	            HeadObjectRequest headObjectRequest = HeadObjectRequest.builder()
+	                .bucket(bucketName)
+	                .key(keyName)
+	                .build();
+
+	            HeadObjectResponse response = s3.headObject(headObjectRequest);
+	            System.out.println("The Amazon S3 object restoration status is "+response.restore());
+
+	        } catch (S3Exception e) {
+	            System.err.println(e.awsErrorDetails().errorMessage());
+	        }
+		
+	}
+
+	public void downloadFiles(String keyName) {
+
+		String home = System.getProperty("user.home");
+
+		System.out.println(home);
+		try {
+            GetObjectRequest objectRequest = GetObjectRequest
+                .builder()
+                .key(keyName)
+                .bucket(bucketName)
+                .build();
+
+            ResponseBytes<GetObjectResponse> objectBytes = s3.getObjectAsBytes(objectRequest);
+            byte[] data = objectBytes.asByteArray();
+
+            // Write the data to a local file.
+            File myFile = new File(home + "/Downloads/" + keyName );
+            OutputStream os = new FileOutputStream(myFile);
+            os.write(data);
+            System.out.println("Successfully obtained bytes from an S3 object");
+            os.close();
+
+        } catch (IOException ex) {
+            ex.printStackTrace();
+        } catch (S3Exception e) {
+            System.err.println(e.awsErrorDetails().errorMessage());
+        }
+		
+		
+		
+		
+		
 
 	}
 
